requirements.txt
anyio==4.4.0
certifi==2024.6.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
keyboard==0.13.5
MouseInfo==0.1.3
pillow==10.4.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.1
python-telegram-bot==21.3
pytweening==1.2.0
sniffio==1.3.1
typing_extensions==4.12.2
